[remo] Rename filenames of ocean and elk remo connectors
This code renames the filenames of ocean and elk remo connectors in
order to easy the management of unit tests
